Move parameter utils
